refactor(agents-desktop): finish shell and IPC modules
Complete the Phase 7-8 split by moving window, menu, state, discovery, and IPC registration domains out of main.ts while preserving channel contracts.
